refactor: route-data hook
